requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.10.0,<2.0.0
langchain==0.1.0
langchain-openai==0.0.5
chromadb==0.4.17
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.3.0
numpy>=1.26.0,<2.0.0